+++ src/expr/sql.rs
@@ -29,7 +29,9 @@ impl<'a> expr::Visitor<'a, ast::Expr> for FromExprVisitor {
             crate::data_type::value::Value::Float(f) => {
                 ast::Expr::Value(ast::Value::Number(format!("{}", **f), false))
             }
-            crate::data_type::value::Value::Text(_) => todo!(),
+            crate::data_type::value::Value::Text(t) => {
+                ast::Expr::Value(ast::Value::SingleQuotedString(format!("{}", **t)))
+            }
             crate::data_type::value::Value::Bytes(_) => todo!(),
             crate::data_type::value::Value::Struct(_) => todo!(),
             crate::data_type::value::Value::Union(_) => todo!(),
@@ -157,11 +159,15 @@ impl<'a> expr::Visitor<'a, ast::Expr> for FromExprVisitor {
             | expr::function::Function::Sin
             | expr::function::Function::Cos
             | expr::function::Function::Sqrt
-            | expr::function::Function::Pow => ast::Expr::Function(ast::Function {
+            | expr::function::Function::Pow
+            | expr::function::Function::Case
+            | expr::function::Function::Md5
+            | expr::function::Function::Concat(_)
+            | expr::function::Function::CharLength
+            | expr::function::Function::Lower
+            | expr::function::Function::Upper => ast::Expr::Function(ast::Function {
                 name: ast::ObjectName(vec![ast::Ident::new(function.to_string())]),
-                args: vec![ast::FunctionArg::Unnamed(ast::FunctionArgExpr::Expr(
-                    arguments[0].clone(),
-                ))],
+                args: arguments.into_iter().map(|e| ast::FunctionArg::Unnamed(ast::FunctionArgExpr::Expr(e))).collect(),
                 over: None,
                 distinct: false,
                 special: false,
@@ -173,6 +179,10 @@ impl<'a> expr::Visitor<'a, ast::Expr> for FromExprVisitor {
                 results: vec![arguments[1].clone()],
                 else_result: Some(Box::new(arguments[2].clone())),
             },
+            expr::function::Function::Position => ast::Expr::Position {
+                expr: arguments[0].clone().into(),
+                r#in: arguments[1].clone().into()
+            }
         }
     }
     // TODO implement this properly
@@ -304,22 +314,70 @@ mod tests {
 
     #[test]
     fn test_from_expr() {
-        let ast_expr: ast::Expr = parse_expr("exp(a*cos(SIN(x) + 2*a + b))").unwrap();
+        let ast_expr: ast::Expr = parse_expr("exp(a*cos(sin(x) + 2*a + b))").unwrap();
         println!("ast::expr = {ast_expr}");
         let expr = Expr::try_from(&ast_expr).unwrap();
         println!("expr = {}", expr);
         let gen_expr = ast::Expr::from(&expr);
         println!("ast::expr = {gen_expr}");
+        assert_eq!(ast_expr, gen_expr)
+    }
+
+    #[test]
+    fn test_from_expr_concat() {
+        let ast_expr: ast::Expr = parse_expr("concat(a, b, c)").unwrap();
+        println!("ast::expr = {ast_expr}");
+        let expr = Expr::try_from(&ast_expr).unwrap();
+        println!("expr = {}", expr);
+        let gen_expr = ast::Expr::from(&expr);
+        println!("ast::expr = {gen_expr}");
+        assert_eq!(ast_expr, gen_expr)
+    }
+
+    #[test]
+    fn test_string_functions() {
+        // Lower
+        let ast_expr: ast::Expr = parse_expr("lower(my_expr)").unwrap();
+        println!("ast::expr = {ast_expr}");
+        let expr = Expr::try_from(&ast_expr).unwrap();
+        println!("expr = {}", expr);
+        let gen_expr = ast::Expr::from(&expr);
+        assert_eq!(ast_expr, gen_expr);
+
+        // Upper
+        let ast_expr: ast::Expr = parse_expr("upper(my_expr)").unwrap();
+        println!("ast::expr = {ast_expr}");
+        let expr = Expr::try_from(&ast_expr).unwrap();
+        println!("expr = {}", expr);
+        let gen_expr = ast::Expr::from(&expr);
+        assert_eq!(ast_expr, gen_expr);
+
+        // CharLength
+        let ast_expr: ast::Expr = parse_expr("char_length(my_expr)").unwrap();
+        println!("ast::expr = {ast_expr}");
+        let expr = Expr::try_from(&ast_expr).unwrap();
+        println!("expr = {}", expr);
+        let gen_expr = ast::Expr::from(&expr);
+        assert_eq!(ast_expr, gen_expr);
+
+        // CharLength
+        let ast_expr: ast::Expr = parse_expr("position('x' in expr)").unwrap();
+        println!("ast::expr = {ast_expr}");
+        let expr = Expr::try_from(&ast_expr).unwrap();
+        println!("expr = {}", expr);
+        let gen_expr = ast::Expr::from(&expr);
+        assert_eq!(ast_expr, gen_expr);
     }
 
     #[test]
     fn test_from_expr_with_var() {
-        let ast_expr: ast::Expr = parse_expr("exp(a*variance(SIN(x) + 2*a + b))").unwrap();
+        let ast_expr: ast::Expr = parse_expr("exp(a*variance(sin(x) + 2*a + b))").unwrap();
         println!("ast::expr = {ast_expr}");
         let expr = Expr::try_from(&ast_expr).unwrap();
         println!("expr = {}", expr);
         let gen_expr = ast::Expr::from(&expr);
         println!("ast::expr = {gen_expr}");
+        assert_eq!(ast_expr, gen_expr)
     }
 
     #[test]