+++ src/sql/expr.rs
@@ -234,6 +234,7 @@ pub trait Visitor<'a, T: Clone> {
         results: Vec<T>,
         else_result: Option<T>,
     ) -> T;
+    fn position(&self, expr: T, r#in: T) -> T;
 }
 
 // For the visitor to be more convenient, we create a few auxiliary objects
@@ -323,7 +324,9 @@ impl<'a, T: Clone, V: Visitor<'a, T>> visitor::Visitor<'a, ast::Expr, T> for V {
             ast::Expr::Extract { field, expr } => todo!(),
             ast::Expr::Ceil { expr, field } => todo!(),
             ast::Expr::Floor { expr, field } => todo!(),
-            ast::Expr::Position { expr, r#in } => todo!(),
+            ast::Expr::Position { expr, r#in } => {
+                self.position(dependencies.get(expr).clone(), dependencies.get(r#in).clone())
+            },
             ast::Expr::Substring {
                 expr,
                 substring_from,
@@ -489,6 +492,8 @@ impl<'a> Visitor<'a, String> for DisplayVisitor {
         };
         case_str.push_str("END");
         case_str
+    fn position(&self, expr: String, r#in: String) -> String {
+        format!("POSITION({} IN {})", expr, r#in)
     }
 }
 
@@ -600,7 +605,7 @@ impl<'a> Visitor<'a, Result<Expr>> for TryIntoExprVisitor<'a> {
                 let x: f64 = FromStr::from_str(number)?;
                 Expr::val(x)
             }
-            ast::Value::SingleQuotedString(_) => todo!(),
+            ast::Value::SingleQuotedString(v) => Expr::val(v.to_string()),
             ast::Value::EscapedStringLiteral(_) => todo!(),
             ast::Value::NationalStringLiteral(_) => todo!(),
             ast::Value::HexStringLiteral(_) => todo!(),
@@ -644,6 +649,12 @@ impl<'a> Visitor<'a, Result<Expr>> for TryIntoExprVisitor<'a> {
             "sqrt" => Expr::sqrt(flat_args[0].clone()),
             "pow" => Expr::pow(flat_args[0].clone(), flat_args[1].clone()),
             "power" => Expr::pow(flat_args[0].clone(), flat_args[1].clone()),
+            "md5" => Expr::md5(flat_args[0].clone()),
+            // string functions
+            "lower" => Expr::lower(flat_args[0].clone()),
+            "upper" => Expr::upper(flat_args[0].clone()),
+            "char_length" => Expr::char_length(flat_args[0].clone()),
+            "concat" => Expr::concat(flat_args.clone()),
             // Aggregates
             "min" => Expr::min(flat_args[0].clone()),
             "max" => Expr::max(flat_args[0].clone()),
@@ -679,6 +690,8 @@ impl<'a> Visitor<'a, Result<Expr>> for TryIntoExprVisitor<'a> {
             case_expr = Expr::case(w.clone(), t.clone(), case_expr.clone());
         }
         Ok(case_expr)
+    fn position(&self, expr: Result<Expr>, r#in: Result<Expr>) -> Result<Expr> {
+        Ok(Expr::position(expr?, r#in?))
     }
 }