+++ CHANGELOG.md
@@ -8,9 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Conversion for `Case` expression [MR1](https://github.com/Qrlew/qrlew/pull/1)
-### Changed
+
+## [0.1.2] - 2023-06-01
+### Added
+- Md5 and Concat functions
+- Protection operation
+- Set operations support
+- String functions `POSITION`, `LOWER`, `UPPER` and `CHAR_LENGTH` [PR9](https://github.com/Qrlew/qrlew/pull/9)
 - Optional handling in or [PR6](https://github.com/Qrlew/qrlew/pull/6)
 
+### Fixed
 ## [0.1.1] - 2023-05-26
 ### Changed
 - Made sqlite optional
@@ -18,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2023-05-25
 ### Added
 - First commit open source
+
 ### Fixed
 ### Changed
 ### Removed